python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bleepstore"
version = "0.1.0"
description = "S3-compatible object storage server"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "pyyaml",
    "prometheus-client",
    "prometheus-fastapi-instrumentator",
    "aiosqlite",
]

[project.optional-dependencies]
aws = [
    "aiobotocore>=2.7.0",
]
gcp = [
    "gcloud-aio-storage>=9.0.0",
]
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "mypy",
    "ruff",
    "aiobotocore>=2.7.0",
    "gcloud-aio-storage>=9.0.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]

[project.scripts]
bleepstore = "bleepstore.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bleepstore"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
